Align middleware tests with module boundaries
Feature-named MiddlewareOptions suites mixed several module contracts and
backend integration scenarios in the root unit-test directories. Keep the
existing per-module convention rather than introducing another test layout.

Place attachment typing and immutability with FunctionSpec and GroupSpec,
equality validation with MiddlewareSpec, and ref propagation with Ref and
Refs. Keep implementation typing and registration rejection in their server
module suites, and run execution order, provider shadowing, and client error
decoding in the existing mock-backend middleware integration suite.

Preserve the existing assertions and type-error coverage, including the
schema type-side versus encoded-input distinction. No production code,
fixtures, or test-discovery configuration changes are needed.
