+++ packages/core/test/GroupSpec.test.ts
@@ -1,5 +1,8 @@
 import { describe, expect, it } from "@effect/vitest";
 import * as GroupSpec from "@confect/core/GroupSpec";
+import * as FunctionSpec from "@confect/core/FunctionSpec";
+import * as MiddlewareSpec from "@confect/core/MiddlewareSpec";
+import * as Schema from "effect/Schema";
 
 describe("isGroupSpec", () => {
   it("checks whether a value is a function spec", () => {
@@ -58,5 +61,98 @@ describe("withName", () => {
 
     expect(renamed.groups).toBe(group.groups);
     expect(renamed.functions).toBe(group.functions);
+  });
+});
+
+describe("middleware options", () => {
+  class AccessDenied extends Schema.TaggedError<AccessDenied>()(
+    "AccessDenied",
+    {},
+  ) {}
+
+  class RequireRole extends MiddlewareSpec.MiddlewareSpec<RequireRole>()(
+    "RequireRole",
+    {
+      options: () =>
+        Schema.Struct({
+          roles: Schema.Array(Schema.Literals(["Internal", "Buyer"])),
+        }),
+      error: () => AccessDenied,
+      functionTypes: { query: true, mutation: true, action: true },
+    },
+  ) {}
+
+  class Observe extends MiddlewareSpec.MiddlewareSpec<Observe>()("Observe", {
+    functionTypes: { query: true, mutation: true, action: true },
+  }) {}
+
+  const query = FunctionSpec.publicQuery({
+    name: "get",
+    returns: () => Schema.String,
+  });
+
+  it("requires exactly the declared attachment options", () => {
+    GroupSpec.make().middleware(RequireRole, { roles: ["Buyer"] });
+    GroupSpec.make().middleware(Observe);
+    // @ts-expect-error
+    GroupSpec.make().middleware(RequireRole);
+    // @ts-expect-error
+    GroupSpec.make().middleware(RequireRole, { roles: ["Unknown"] });
+    // @ts-expect-error
+    GroupSpec.make().middleware(RequireRole, undefined);
+    // @ts-expect-error
+    GroupSpec.make().middleware(Observe, {});
+  });
+
+  it("accepts the options schema's type rather than its encoded input", () => {
+    class Limit extends MiddlewareSpec.MiddlewareSpec<Limit>()("Limit", {
+      options: () => Schema.Struct({ limit: Schema.FiniteFromString }),
+      functionTypes: { query: true, mutation: false, action: false },
+    }) {}
+    const group = GroupSpec.make().middleware(Limit, { limit: 5 });
+    expect(group.middlewareAttachments[0]?.options).toEqual({ limit: 5 });
+    // @ts-expect-error
+    GroupSpec.make().middleware(Limit, { limit: "5" });
+  });
+
+  it("allows different options at both attachment levels and across their boundary", () => {
+    const internal = query.middleware(RequireRole, { roles: ["Internal"] });
+    const group = GroupSpec.make().middleware(RequireRole, {
+      roles: ["Buyer"],
+    });
+    GroupSpec.validateMiddleware(
+      GroupSpec.make().addFunction(
+        internal.middleware(RequireRole, { roles: ["Buyer"] }),
+      ),
+    );
+    GroupSpec.validateMiddleware(
+      group.middleware(RequireRole, { roles: ["Internal"] }).addFunction(query),
+    );
+    GroupSpec.validateMiddleware(group.addFunction(internal));
+    GroupSpec.validateMiddleware(
+      GroupSpec.make()
+        .addFunction(internal)
+        .middleware(RequireRole, { roles: ["Buyer"] }),
+    );
+  });
+
+  it("rejects equivalent instances only during validation, including group/function overlaps", () => {
+    const first = query.middleware(RequireRole, { roles: ["Internal"] });
+    const second = first.middleware(RequireRole, { roles: ["Internal"] });
+    const group = GroupSpec.make().middleware(RequireRole, {
+      roles: ["Internal"],
+    });
+    for (const candidate of [
+      GroupSpec.make().addFunction(second),
+      group.middleware(RequireRole, { roles: ["Internal"] }),
+      group.addFunction(first),
+      GroupSpec.make()
+        .addFunction(first)
+        .middleware(RequireRole, { roles: ["Internal"] }),
+    ]) {
+      expect(() => GroupSpec.validateMiddleware(candidate)).toThrowError(
+        /RequireRole.*equivalent options.*1 and 2/,
+      );
+    }
   });
 });
