+++ packages/server/test/RegisteredFunctions.test.ts
@@ -1,6 +1,16 @@
-import { RegisteredConvexFunction, RegisteredFunctions } from "@confect/server";
+import { FunctionSpec, GroupSpec, MiddlewareSpec } from "@confect/core";
+import {
+  FunctionImpl,
+  GroupImpl,
+  MiddlewareImpl,
+  RegisteredConvexFunction,
+  RegisteredFunctions,
+} from "@confect/server";
 import { describe, expect, expectTypeOf, it } from "@effect/vitest";
 import type { RegisteredQuery } from "convex/server";
+import * as Effect from "effect/Effect";
+import * as Layer from "effect/Layer";
+import * as Schema from "effect/Schema";
 import databaseSchema from "./mock-backend/fixtures/confect/_generated/schema";
 import registeredFunctions from "./mock-backend/fixtures/confect/_generated/registeredFunctions/groups/notes";
 // Imported as a value (not `import type`) because its type parameterizes
@@ -28,6 +38,42 @@ describe("Registered functions", () => {
 });
 
 describe("buildForGroup", () => {
+  it("rejects equivalent options during registration without requiring codegen", () => {
+    class GroupPolicy extends MiddlewareSpec.MiddlewareSpec<GroupPolicy>()(
+      "GroupPolicy",
+      {
+        options: () => Schema.Struct({ label: Schema.String }),
+        functionTypes: { query: true, mutation: true, action: true },
+      },
+    ) {}
+
+    const query = FunctionSpec.publicQuery({
+      name: "get",
+      returns: () => Schema.String,
+    }).middleware(GroupPolicy, { label: "same" });
+    const group = GroupSpec.make()
+      .middleware(GroupPolicy, { label: "same" })
+      .addFunction(query);
+    const layer = GroupImpl.make(databaseSchema, group).pipe(
+      Layer.provide(
+        FunctionImpl.make(databaseSchema, group, "get", () =>
+          Effect.succeed("ok"),
+        ),
+      ),
+      Layer.provide(
+        MiddlewareImpl.make(databaseSchema, GroupPolicy, (effect) => effect),
+      ),
+      GroupImpl.finalize,
+    );
+    expect(() =>
+      RegisteredFunctions.buildForGroup<typeof group>(
+        databaseSchema,
+        layer,
+        RegisteredConvexFunction.make,
+      ),
+    ).toThrowError(/GroupPolicy.*equivalent options.*function "get"/);
+  });
+
   it("registers only the requested group", () => {
     const registered = RegisteredFunctions.buildForGroup<typeof notesSpec>(
       databaseSchema,
